requirement.txt
dnspython==2.7.0
pymongo==4.10.1
PyYAML==6.0.2
